core/serializers.py
from rest_framework.serializers import ModelSerializer
from .models import Collection

class CollectionSerializer(ModelSerializer):
    class Meta:
        model = Collection
        fields = ['id', 'name']
